Training code added. IT IS UNTESTED

+++ i915_controller.h
@@ -69,7 +69,7 @@ typedef struct {
 
     UINT64 (*read64)(UINT64 reg);
     struct {
-        UINT8 Port;
+        UINT32 Port;
         UINT8 DDI;
         UINT8 Plane;
         UINT8 Transcoder;
